final fixes
