+++ app/src/foss/AndroidManifest.xml
@@ -1,0 +1,9 @@
+<?xml version="1.0" encoding="utf-8"?>
+<manifest
+    xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:tools="http://schemas.android.com/tools"
+    package="com.gh4a">
+    <application>
+        <meta-data android:name="com.crashlytics.ApiKey" tools:node="remove"/>
+    </application>
+</manifest>
