--- app/src/foss/java/com/gh4a/FabricApplication.java
@@ -1,12 +1,0 @@
-package com.gh4a;
-
-import android.app.Application;
-
-public class FabricApplication {
-    public static void onCreate(Application app) {
-
-    }
-    public static void trackVisitedUrl(String url, int position) {
-
-    }
-}
