+++ app/src/foss/java/com/gh4a/utils/CrashReportingHelper.java
@@ -1,0 +1,11 @@
+package com.gh4a.utils;
+
+import android.app.Application;
+
+public class CrashReportingHelper {
+
+    public static void onCreate(Application app) {
+    }
+    public static void trackVisitedUrl(Application app, String url) {
+    }
+}
